Update the non-zero check for `bits`.
Looks like there was an update for the so -> So migration, but an `oh`
in the default tactics got overlooked. Should've been `Oh`.
Someone reported an error with this:
http://stackoverflow.com/questions/29425769/cant-create-simple-binary-data-structure-with-idrisnet2

Replaced the So > 0 impl with GT 0. And used `auto` instead of the
`default tactics`. Shouldn't run into the case where something compiles
but doesn't work on "normal" invocation again.